ref: Refactored parsing test project
